code/detection/requirements.txt
# --------------------- If use pip -----------------------------
numpy==1.17
opencv-python
torch
torchvision
onnx
matplotlib
tqdm
pillow
scikit-image
pycocotools

# -------------------------------------------------------------- #
#                          optional                              #
# -------------------------------------------------------------- #

# ---------- Nvidia Apex for mixed precision training ----------
# git clone https://github.com/NVIDIA/apex && cd apex && pip install -v --no-cache-dir --global-option="--cpp_ext" --global-option="--cuda_ext" . --user && cd .. && rm -rf apex

# ------------------- Tensorboard ------------------------------
# tb-nightly
# future

# ------------------ If use anconda -----------------------------
# conda update -yn base -c defaults conda
# conda install -yc anaconda numpy opencv matplotlib tqdm pillow ipython future
# conda install -yc conda-forge scikit-image pycocotools tensorboard
# conda install -yc spyder-ide spyder-line-profiler
# conda install -yc pytorch pytorch torchvision
# conda install -yc conda-forge protobuf numpy && pip install onnx  # https://github.com/onnx/onnx#linux-and-macos
